- moved DB adapters to infrastructure
