+++ infrastructure/postgres/user/database_user_query_adapter.py
@@ -4,7 +4,7 @@ from uuid import UUID
 from sqlalchemy.engine import Engine
 from sqlalchemy.orm import Session
 
-from domain.user.adapter.database.utils.user_utils import entity_to_user, get_user_entity
+from infrastructure.postgres.user.utils.user_utils import entity_to_user, get_user_entity
 from domain.user.exceptions import UserNotFound
 from domain.user.model.User import User
 from domain.user.user_query_port import UserQueryPort
