+++ tests/conftest.py
@@ -27,4 +27,4 @@ def db_engine() -> Engine:
 @pytest.fixture(scope="session")
 def user_config() -> UserConfig:
     return UserConfig(username_char_wl="qwertyuiopasdfghjklzxcvbnmQWERTYUIOPASDFGHJKLZXCVBNM1234567890.:_-+=",
-                        username_min_len=4, username_max_len=16, passwd_min_len=7, passwd_max_len=25)
+                      username_min_len=4, username_max_len=16, passwd_min_len=7, passwd_max_len=25)
