+++ tests/infrastructure/postgres/user/test_modify_change_password_it.py
@@ -4,9 +4,9 @@ import pytest
 from argon2 import PasswordHasher
 
 from domain.config.model.UserConfig import UserConfig
-from domain.user.adapter.database.database_user_modify_adapter import DatabaseUserModifyAdapter
-from domain.user.adapter.database.database_user_query_adapter import DatabaseUserQueryAdapter
-from domain.user.exceptions import UserNotFound, PasswordSyntaxError, ChangePasswordError
+from infrastructure.postgres.user.database_user_modify_adapter import DatabaseUserModifyAdapter
+from infrastructure.postgres.user.database_user_query_adapter import DatabaseUserQueryAdapter
+from domain.user.exceptions import PasswordSyntaxError, ChangePasswordError
 from domain.user.model.ChangePasswordCmd import ChangePasswordCmd
 from domain.user.model.UserCreateCmd import UserCreateCmd
 from tests.decors import using_database
