+++ tests/infrastructure/postgres/user/test_modify_delete_user_it.py
@@ -1,7 +1,7 @@
 import pytest
 
-from domain.user.adapter.database.database_user_modify_adapter import DatabaseUserModifyAdapter
-from domain.user.adapter.database.database_user_query_adapter import DatabaseUserQueryAdapter
+from infrastructure.postgres.user.database_user_modify_adapter import DatabaseUserModifyAdapter
+from infrastructure.postgres.user.database_user_query_adapter import DatabaseUserQueryAdapter
 from domain.user.exceptions import UserNotFound, UserDeleteError
 from domain.user.model.UserCreateCmd import UserCreateCmd
 from domain.user.model.UserDeleteCmd import UserDeleteCmd
