+++ tests/infrastructure/postgres/user/test_query_find_user_it.py
@@ -3,8 +3,8 @@ from uuid import UUID
 import pytest
 from argon2 import PasswordHasher
 
-from domain.user.adapter.database.database_user_modify_adapter import DatabaseUserModifyAdapter
-from domain.user.adapter.database.database_user_query_adapter import DatabaseUserQueryAdapter
+from infrastructure.postgres.user.database_user_modify_adapter import DatabaseUserModifyAdapter
+from infrastructure.postgres.user.database_user_query_adapter import DatabaseUserQueryAdapter
 from domain.user.exceptions import UserNotFound
 from domain.user.model.UserCreateCmd import UserCreateCmd
 from tests.decors import using_database
